+++ radio/src/gui/128x64/model_telemetry.cpp
@@ -166,7 +166,7 @@ void menuModelSensor(event_t event)
 
   int8_t sub = menuVerticalPosition;
 
-  for (uint8_t i=0; i<LCD_LINES-1; i++) {
+  for (uint32_t i=0; i<LCD_LINES-1; i++) {
     coord_t y = MENU_HEADER_HEIGHT + 1 + i*FH;
     uint8_t k = i + menuVerticalOffset;
 
@@ -438,7 +438,7 @@ void menuModelTelemetryFrsky(event_t event)
       break;
   }
 
-  for (uint8_t i=0; i<LCD_LINES-1; i++) {
+  for (uint32_t i=0; i<LCD_LINES-1; i++) {
     coord_t y = MENU_HEADER_HEIGHT + 1 + i*FH;
     uint8_t k = i + menuVerticalOffset;
     for (int j=0; j<=k; j++) {
