+++ radio/src/gui/212x64/model_select.cpp
@@ -238,7 +238,7 @@ void menuModelSelect(event_t event)
 
   TITLE(STR_MENUMODELSEL);
 
-  for (uint8_t i=0; i<NUM_BODY_LINES; i++) {
+  for (uint32_t i=0; i<NUM_BODY_LINES; i++) {
     coord_t y = MENU_HEADER_HEIGHT + 1 + i*FH;
     uint8_t k = i+menuVerticalOffset;
