+++ radio/src/gui/212x64/radio_diaganas.cpp
@@ -24,7 +24,7 @@ void menuRadioDiagAnalogs(event_t event)
 {
   SIMPLE_SUBMENU(STR_MENU_RADIO_ANALOGS, 0);
 
-  for (uint8_t i=0; i<NUM_STICKS+NUM_POTS+NUM_SLIDERS; i++) {
+  for (uint32_t i=0; i<NUM_STICKS+NUM_POTS+NUM_SLIDERS; i++) {
     coord_t y = MENU_HEADER_HEIGHT + 1 + (i/2)*FH;
     uint8_t x = i&1 ? LCD_W/2 + FW : 0;
     lcdDrawNumber(x, y, i+1, LEADING0|LEFT, 2);
